refactor(tests): move pool tests to tests/core/pool/
Consolidate pool config, health, and pool_health tests into
tests/core/pool/ subdirectory. Make PoolConfig fields and
ConnectionMeta pub for external test access.
