+++ crates/sentinel-driver/src/pool/health.rs
@@ -42,7 +42,7 @@ pub enum HealthCheckStrategy {
 
 /// Metadata for a pooled connection — used for idle timeout and max lifetime.
 #[derive(Debug)]
-pub(crate) struct ConnectionMeta {
+pub struct ConnectionMeta {
     /// When this connection was created.
     pub created_at: Instant,
     /// When this connection was last returned to the pool.
@@ -74,40 +74,5 @@ impl ConnectionMeta {
     /// Check if the connection has exceeded its max lifetime.
     pub fn is_lifetime_expired(&self, max_lifetime: std::time::Duration) -> bool {
         self.created_at.elapsed() > max_lifetime
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use std::time::Duration;
-
-    #[test]
-    fn test_connection_meta_new() {
-        let meta = ConnectionMeta::new();
-        assert!(!meta.is_broken);
-        assert!(meta.created_at.elapsed() < Duration::from_secs(1));
-    }
-
-    #[test]
-    fn test_connection_meta_touch() {
-        let mut meta = ConnectionMeta::new();
-        let first_used = meta.last_used;
-        // Touch updates last_used
-        std::thread::sleep(Duration::from_millis(1));
-        meta.touch();
-        assert!(meta.last_used >= first_used);
-    }
-
-    #[test]
-    fn test_idle_not_expired() {
-        let meta = ConnectionMeta::new();
-        assert!(!meta.is_idle_expired(Duration::from_secs(600)));
-    }
-
-    #[test]
-    fn test_lifetime_not_expired() {
-        let meta = ConnectionMeta::new();
-        assert!(!meta.is_lifetime_expired(Duration::from_secs(3600)));
     }
 }
