+++ tests/core/pool/mod.rs
@@ -1,0 +1,2 @@
+mod config;
+mod health;
